// src/services/use-cases/subjects/utils/enums/general-information-titles.enums.ts
export enum GeneralInfoTitles {
  MASTERING = 'Mestring',
  MOTIVATION = 'Motivation',
  RESOURCES = 'Resources',
  ROLLS = 'Roller',
  HABITS = 'Vaner',
  EDUCATION_OR_JOB = 'Uddannelse/Job',
  LIFE_STORY = 'Livshistorie',
  SOCIAL_CIRCLE = 'Netværk',
  HEALTH_INFORMATION = 'Helbredsoplysninger',
  ASSISTIVE_DEVICES = 'Hjælpemidler',
  HOME_INTERIOR_DESIGN = 'Boligens indretning',
}

export enum GeneralInfoDescriptions {
  MASTERING_DESCRIPTION = 'Borgerens bevidste eller ubevidste håndtering af livet/sygdommen - både udfordringer og muligheder.',
  MOTIVATION_DESCRIPTION = 'Drivkraften bag at borgeren handler på en bestemt måde eller går i gang med/opretholder en opgave/indsats',
  RESOURCES_DESCRIPTION = 'De fysiske eller mentale kræft, som borgeren i et vist omfang har til rådighed og kan udnytte. Fysiske kræfter kan fx være i form af fysisk sundhed og styrke. Mentale kræfter kan fx være i form af psykisk sundhed og styrke, herunder tanker og måder at forholde sig til situationer på og andre mennesker på.',
  ROLES_DESCRIPTION = 'De roller som er særligt vigtige for borgeren i forhold til familie, arbejde og samfund.',
  HABITS_DESCRIPTION = 'Regelmæssig adfærd som borgeren har tillært gennem stadig gentagelse og udførelse helt eller delvist ubevidst. Vaner er fx døgnrytmen, måden at blive tiltalt på, kontakt med medmennesker og relationer, måde at anskue verden på.',
  EDUCATION_OR_JOB_DESCRIPTION = 'Nuværende eller tidligere uddannelses- og/eller erhvervsmæssig baggrund.',
  LIFE_STORY_DESCRIPTION = 'En beskrivelse af borgerens oplevelse af væsentlige begivenheder, interesser og gøremål igennem livet.',
  SOCIAL_CIRCLE_DESCRIPTION = 'Personer som er tæt på borgeren, og som giver praktisk og/eller følelsesmæssigt støtte og omsorg overfor borgeren. Netværk kan være offentlig eller privat. Et offentligt netværk består af personlige hjælpere, sunhedspersonale og andre professionelle primære omsorgsgivere. Privat netværk af familie, slægtning, venner og bekendtskaber.',
  HEALTH_INFORMATION_DESCRIPTION = 'Aktuelle eller tidligere sygdomme og handicap der har betydning for borgerens situation. Sundhedsfaglige kontakter - medarbejdere eller enheder indenfor sundhedsvæsenet borgeren er tilknyttet, fx øjenlæge, tandlæge, fodterapeut eller afdeling/ambulatorium.',
  ASSISTIVE_DEVICES_DESCRIPTION = 'Udstyr, produkter og teknologi som anvendes af borgeren i daglige aktiviteter, inkl. såddanne som er tilpasset eller særligt fremstillet til, implanteret i, plcaeret på eller nær personen, som anvender dem. (Inklusiv almindelige genstande og hjælpemidler og teknologi til personlig anvendelse.',
  HOME_INTERIOR_DESIGN_DESCRIPTION = 'En beksrivelse af boligens fysiske rammer og omgivelser, der har betydning for borgerens hverdagsliv og funktionsevne.',
}