// src/services/use-cases/subjects/utils/enums/function-ability-titles.enums.ts
export enum FunctionAbilityTitles {
  SELF_CARE = 'Egenomsorg',
  PRACTICAL_TASKS = 'Praktiske opgaver',
  MOBILITY = 'Mobilitet',
  MENTAL_FUNCTIONS = 'Mentale funktioner',
  SOCIETY = 'Samfundsliv',
}

export enum FunctionAbilitySubTitles {
  /*SELF_CARE*/
  SELF_CARE_BATHING = 'Vaske sig',
  SELF_CARE_BODY_CARE = 'Kropspleje',
  SELF_CARE_UNDRESSING_AND_DRESSING = 'Af- og påklædning',
  SELF_CARE_DRINKING = 'Drikke',
  SELF_CARE_EATING = 'Spise',
  SELF_CARE_FOOD_INTAKE = 'Fødeindtagelse',
  SELF_CARE_HEALTH_CARE = 'Varetage egen sundhed',
  SELF_CARE_TOILET = 'Gå på toilet',

  /*PRACTICAL_TASKS*/
  PRACTICAL_TASKS_HOUSEWORK = 'Lave husligt arbejde',
  PRACTICAL_TASKS_COOKING = 'Lave mad',
  PRACTICAL_TASKS_DAILY_ROUTINES = 'Udføre daglige rutiner',
  PRACTICAL_TASKS_GETTING_GROCERIES_AND_SERVICES = 'Skaffe sig varer og tjenesteydelser',

  /*MOBILITY*/
  MOBILITY_LIFT_AND_CARRYING = 'Løfte og bære',
  MOBILITY_MOVE_AROUND = 'Bevæge sig rundt',
  MOBILITY_USE_OF_TRANSPORTATION = 'Bruge transportmidler',
  MOBILITY_MOVEMENT_IN_DIFFERENT_ENVIRONMENTS = 'Færden i forskellige omgivelser',
  MOBILITY_TO_MOVE = 'Forflytte sig',
  MOBILITY_CHANGE_BODY_POSITION = 'Ændre kropsstilling',
  MOBILITY_STRENGTH = 'Muskelstyrke',
  MOBILITY_WALK = 'Gå',
  MOBILITY_ENDURANCE = 'Udholdenhed',

  /*MENTAL_FUNCTIONS*/
  MENTAL_FUNCTIONS_USE_OF_COMMUNICATION_DEVICES = 'Anvende kommunikationsudstyr og teknikker',
  MENTAL_FUNCTIONS_MEMORY = 'Hukommelse',
  MENTAL_FUNCTIONS_ORIENTATION = 'Orienteringsevne',
  MENTAL_FUNCTIONS_COGNITIVE_FUNCTIONS = 'Overordnede kognitive funktioner',
  MENTAL_FUNCTIONS_EMOTIONAL_FUNCTIONS = 'Følelsesfunktioner',
  MENTAL_FUNCTIONS_ENERGY_AND_POWER_OF_ACTION = 'Energi og handlekraft',
  MENTAL_FUNCTIONS_LEARN_NEW_SKILLS = 'Tilegne sig færdigheder',
  MENTAL_FUNCTIONS_PROBLEM_SOLVING = 'Problemløsning',

  /*SOCIETY*/
  SOCIETY_JOB = 'Have lønnet beskæftigelse',
}